backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-cloud-pubsub>=2.18.0
google-cloud-firestore>=2.11.0
google-cloud-bigquery>=3.11.0
google-cloud-aiplatform>=1.30.0
google-cloud-logging>=3.8.0
google-cloud-error-reporting>=1.10.0
firebase-admin>=6.2.0
PyJWT>=2.8.0
redis>=4.6.0
pytest>=7.4.0
httpx>=0.24.1
google-cloud-secret-manager>=2.16.0
python-multipart>=0.0.6
locust>=2.19.0
